dev/requirements.prod.txt
# Production dependencies for Notion-Hugo pipeline
notion-client>=2.0.0
pyyaml>=6.0
requests>=2.28.0
python-frontmatter>=1.0.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
fs>=2.4.0
tabulate>=0.9.0